# notes/App/Note.API/View/Auth/ViewLoginModel.cs
namespace Note.API.View.Auth
{
    public class ViewLoginModel
    {
        //TODO: добавить серверную валидацию полей 
        public string Password { get; set; }
        public string Email { get; set; }
    }
}